Create setup.sh
In this updated version, we have added the INSTALL_DIR variable, which defines the target installation directory (/opt/DebianSystemDiff in this case). The script creates the installation directory if it doesn't exist and then copies all the files and folders from the current location (the downloaded repository) to the installation directory.

The dependencies are installed using pip3 from the requirements.txt file located in the installation directory.

Please make sure to update the installation directory path (INSTALL_DIR) if you prefer a different location.

After updating the setup.sh file, commit and push it to your GitHub repository with an appropriate comment describing the changes made.